+++ include/murmurhash.h
@@ -8,6 +8,6 @@
 
 #include <stdint.h>
 
-void MurmurHash3_x64_128(const void *key, const int len, const uint32_t seed, void *out);
+void MurmurHash3_x64_128(const void *key, const size_t len, const uint32_t seed, void *out);
 
 #endif  // MURMURHASH_H_
